package/frr: update to 9.1.1
Update to the latest 9.x version. Protobuf is now mandatory.

Thanks to the suggestions from Thomas Petazzoni: Since there is
a dependency on BR2_PACKAGE_PROTOBUF_C, the dependencies of protobuf_c
should be imported into FRR's Config.in and the comment should be
aligned accordingly.
Note that in order to enable FRR, the following must be enabled
to support C++:
 BR2_TOOLCHAIN_BUILDROOT_CXX=y
 BR2_PACKAGE_FRR=y

COPYING hash updated:
The COPYING file has evolved to expose more details: the respective
licenses are now listed, except for the GPLv3 license. The GPLv3
license is only required when FRR is compiled with the GCC plugin
that extends printf type checks. This plugin should not be enabled
for production code.
For more information, see FRR's workflow.rst document.

Signed-off-by: Vincent Jardin <vjardin@free.fr>
Suggested-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>

frr/Config.in: enable CXX